map-coloring/backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
pillow>=9.0.0
opencv-python-headless>=4.7.0
numpy>=1.23.0
scikit-image>=0.19.0
networkx>=2.8.0
pydantic>=2.0.0
